Crate sunrise_time

Time Service

This service takes care of anything related with time.

Modules

timezone
TimeZone module

Structs

RTCManager
RTC interface.

Rtc
IBM Real Time Clock provides access to the current date and time (at second precision). The Real Time Clock is actually part of the CMOS on usual IBM/PC setups.

StaticService
Entry point interface.

Statics

CAPABILITIES

HEADER

RTC_INSTANCE
Global instance of Rtc.

Functions

main

update_rtc
Task responsible for updating the RTC_INSTANCE's current time every second.