Struct sunrise_time::timezone::TimeZoneManager

pub struct TimeZoneManager {
    location: LocationName,
    my_rules: TimeZoneRule,
    temp_rules: TimeZoneRule,
}

Global instance handling I/O and storage of the device rules.

Fields
location: LocationName

The location name of this device.

my_rules: TimeZoneRule

Rules of this device.

temp_rules: TimeZoneRule

Temporary rules storage used during timezone conversion.

Implementations

impl TimeZoneManager

pub fn get_device_location_name(&self) -> LocationName

Get the time zone name used on this devie.

pub fn set_device_location_name(
&mut self,
location: LocationName
) -> Result<(), Error>

Set the time zone name used on this devie.

Note:

This also load the new timezone rule.

pub fn set_device_location_name_unchecked(&mut self, location: LocationName)

Set the time zone name used on this devie.

pub fn get_total_location_name_count(&self) -> Result<u32, Error>

Get the total count of location name available

pub fn load_timezone_rule(
&mut self,
location: LocationName,
timezone_rule: Option<&mut TimeZoneRule>
) -> Result<(), Error>

Load a time zone rule.

pub fn get_my_rules(&self) -> &TimeZoneRule

Get the device timezone rule.